Added extra screen for signing up to competition
