+++ Schutter/urls.py
@@ -5,37 +5,42 @@
 #  Licensed under BSD-3-Clause-Clear. See LICENSE file for details.
 
 from django.urls import path
-from . import views
+from . import view_registreer_nhb, view_profiel, view_voorkeuren, view_inschrijven_uitschrijven, view_leeftijdsklassen
+
 
 app_name = 'Schutter'
 
 urlpatterns = [
     path('',
-         views.ProfielView.as_view(),
+         view_profiel.ProfielView.as_view(),
          name='profiel'),
 
     path('registreer/',
-         views.RegistreerNhbNummerView.as_view(),
+         view_registreer_nhb.RegistreerNhbNummerView.as_view(),
          name='registreer'),
 
     path('voorkeuren/<nhblid_pk>/',
-         views.VoorkeurenView.as_view(),
+         view_voorkeuren.VoorkeurenView.as_view(),
          name='voorkeuren-nhblid'),
 
     path('voorkeuren/',
-         views.VoorkeurenView.as_view(),
+         view_voorkeuren.VoorkeurenView.as_view(),
          name='voorkeuren'),
 
     path('leeftijdsklassen/',
-         views.LeeftijdsklassenView.as_view(),
+         view_leeftijdsklassen.LeeftijdsklassenView.as_view(),
          name='leeftijdsklassen'),
 
+    path('regiocompetitie/inschrijven/<deelcomp_pk>/<schutterboog_pk>/bevestig/',
+         view_inschrijven_uitschrijven.RegiocompetitieInschrijvenBevestigView.as_view(),
+         name='bevestig-inschrijven'),
+
     path('regiocompetitie/inschrijven/<deelcomp_pk>/<schutterboog_pk>/',
-         views.RegiocompetitieInschrijvenView.as_view(),
+         view_inschrijven_uitschrijven.RegiocompetitieInschrijvenView.as_view(),
          name='inschrijven'),
 
     path('regiocompetitie/uitschrijven/<regiocomp_pk>/',
-         views.RegiocompetitieUitschrijvenView.as_view(),
+         view_inschrijven_uitschrijven.RegiocompetitieUitschrijvenView.as_view(),
          name='uitschrijven'),
 ]
